Create annotations from actual data.

+++ telcoinc/Fixtures/GeolocationFixture.swift
@@ -9,11 +9,11 @@
 import Foundation
 
 public class GeolocationFixture {
-    private static let LATITUDE_LOWER_BOUND = Decimal(41.3838005695);
-    private static let LATITUDE_UPPER_BOUND = Decimal(41.4468805643);
+    private static let LATITUDE_LOWER_BOUND = Double(41.3838005695);
+    private static let LATITUDE_UPPER_BOUND = Double(41.4468805643);
     
-    private static let LONGITUDE_LOWER_BOUND = Decimal(2.1400831772);
-    private static let LONGITUDE_UPPER_BOUND = Decimal(2.1970747544);
+    private static let LONGITUDE_LOWER_BOUND = Double(2.1400831772);
+    private static let LONGITUDE_UPPER_BOUND = Double(2.1970747544);
     
     
     public static func create() -> Geolocation {
@@ -23,7 +23,7 @@ public class GeolocationFixture {
         return Geolocation(latitude: latitude, longitude: longitude)
     }
     
-    private static func randomBetweenNumbers(firstNum: Decimal, secondNum: Decimal) -> Decimal {
-        return Decimal(arc4random()) / Decimal(UINT32_MAX) * abs(firstNum - secondNum) + min(firstNum, secondNum)
+    private static func randomBetweenNumbers(firstNum: Double, secondNum: Double) -> Double {
+        return Double(arc4random()) / Double(UINT32_MAX) * abs(firstNum - secondNum) + min(firstNum, secondNum)
     }
 }
